interface: refactored chat window into header, messages, and window
