+++ pkg/interface/src/lib/util.js
@@ -275,3 +275,38 @@ export function stringToSymbol(str) {
   }
   return result;
 }
+
+export function scrollIsAtTop(container) {
+  if (
+    (navigator.userAgent.includes("Safari") &&
+      navigator.userAgent.includes("Chrome")) ||
+    navigator.userAgent.includes("Firefox")
+  ) {
+    return container.scrollTop === 0;
+  } else if (navigator.userAgent.includes("Safari")) {
+    return (
+      container.scrollHeight + Math.round(container.scrollTop) <=
+      container.clientHeight + 10
+    );
+  } else {
+    return false;
+  }
+}
+
+export function scrollIsAtBottom(container) {
+  if (
+    (navigator.userAgent.includes("Safari") &&
+      navigator.userAgent.includes("Chrome")) ||
+    navigator.userAgent.includes("Firefox")
+  ) {
+    return (
+      container.scrollHeight - Math.round(container.scrollTop) <=
+      container.clientHeight + 10
+    );
+  } else if (navigator.userAgent.includes("Safari")) {
+    return container.scrollTop === 0;
+  } else {
+    return false;
+  }
+}
+
